diff abundance filter  modify

+++ script/05_filter_abundance.py
@@ -14,7 +14,7 @@ def read_params(args):
                         help="set the marker file in")
     parser.add_argument('-o', '--outdir', dest='outdir', metavar='DIR', type=str, required=True,
                         help="set the work dir")
-    parser.add_argument('--cut_off', dest='cut_off', metavar='cut_off', type=int, required=True,
+    parser.add_argument('--cut_off', dest='cut_off', metavar='cut_off', type=float, required=True,
                         help="set the group file")
     parser.add_argument('--quantile', dest='quantile', metavar='quantile', type=float, default=0.25,
                         help="thr sum of raw  set quantile;0.25 or 0.5 or 0.75")
